requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
pillow==11.3.0
proglog==0.1.12
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.15.0
yt-dlp==2025.6.30
pydantic>=1.10.12,<2.0.0
fastapi>=0.104.0,<0.105.0